created popover for user settings
