+++ client/src/components/temp.js
@@ -1,0 +1,39 @@
+import * as React from 'react';
+import Popover from '@mui/material/Popover';
+import Typography from '@mui/material/Typography';
+import Button from '@mui/material/Button';
+
+export default function BasicPopover() {
+    const [anchorEl, setAnchorEl] = React.useState(null);
+
+    const handleClickPopper = (event) => {
+        setAnchorEl(event.currentTarget);
+    };
+
+    const handleClosePopper = () => {
+        setAnchorEl(null);
+    };
+
+    const openPopper = Boolean(anchorEl);
+    const id = openPopper ? 'simple-popover' : undefined;
+
+    return (
+        <div>
+        <Button aria-describedby={id} variant="contained" onClick={handleClickPopper}>
+            Open Popover
+        </Button>
+        <Popover
+            id={id}
+            open={openPopper}
+            anchorEl={anchorEl}
+            onClose={handleClosePopper}
+            anchorOrigin={{
+            vertical: 'bottom',
+            horizontal: 'left',
+            }}
+        >
+            <Typography sx={{ p: 2 }}>The content of the Popover.</Typography>
+        </Popover>
+        </div>
+    );
+}
